+++ helloworld-blockchain-node/src/main/java/com/xingkaichun/helloworldblockchain/node/dto/blockchainbranch/BlockchainBranchBlockDto.java
@@ -3,9 +3,11 @@ package com.xingkaichun.helloworldblockchain.node.dto.blockchainbranch;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class BlockchainBranchBlockDto {
 
-    private int blockHeight;
+    private BigInteger blockHeight;
     private String blockHash;
 }
