+++ helloworld-blockchain-node/src/main/java/com/xingkaichun/helloworldblockchain/node/model/BlockchainBranchBlockEntity.java
@@ -3,9 +3,11 @@ package com.xingkaichun.helloworldblockchain.node.model;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class BlockchainBranchBlockEntity {
 
-    private int blockHeight;
+    private BigInteger blockHeight;
     private String blockHash;
 }
